requirements.txt
Django==5.1.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.5.0
psycopg[binary]==3.2.3
python-decouple==3.8
celery==5.4.0
redis==5.2.0

